// src/app/models/esyaModels/esyaYukDto.ts
export interface EsyaYukDto {
  Id: number;
  SeferId: number;
  YukId: number;
  FirmaYukNo: string;
  TasimaTuruKodu: number;
  TasimaTuru: string;
  GonderenVergiNo: string;
  GonderenUnvan: string;
  AliciVergiNo: string;
  AliciUnvan: string;
  YuklemeUlkeKodu: string;
  YuklemeIlMernisKodu: number;
  YuklemeIlMernisAdi: string;
  YuklemeIlceMernisKodu: number;
  YuklemeIlceMernisAdi: string;
  BosaltmaUlkeKodu: string;
  BosaltmaIlMernisKodu: number;
  BosaltmaIlMernisAdi: string;
  BosaltmaIlceMernisKodu: number;
  BosaltmaIlceMernisAdi: string;
  YuklemeTarihi: string;
  YuklemeSaati: string;
  BosaltmaTarihi: string;
  BosaltmaSaati: string;
  YukCinsId: number;
  YukCinsAdi: string;
  YukCinsDigerAciklama: string;
  YukMiktariBirimi: string;
  YukMiktariBirimAdi: string;
  YukMiktari: string;
  TehlikeliMaddeTasimaKodu: number;
  TehlikeliMaddeTasimaSekli: string;
  UnId: string;
  MuafiyetTuru: string;
  AktifIptal: string;
  IptalKodu: string;
  IptalAciklama: string;
}